Modificación run usuario

+++ src/empresa/entity/Cliente.java
@@ -20,7 +20,7 @@ public class Cliente extends Usuario {
     }
 
     //Constructor con atributos de la clase Cliente y atributos heredados de la clase Usuario
-    public Cliente(String nombreUsuario, String apellido1, String apellido2, LocalDate fechaDeNacimiento, Long run, String razonSocial, String giroEmpresa, Long rut, String telefonoRepresentante, String direccion, String comuna) {
+    public Cliente(String nombreUsuario, String apellido1, String apellido2, LocalDate fechaDeNacimiento, int run, String razonSocial, String giroEmpresa, Long rut, String telefonoRepresentante, String direccion, String comuna) {
         super(nombreUsuario, apellido1, apellido2, fechaDeNacimiento, run);
         this.razonSocial = razonSocial;
         this.giroEmpresa = giroEmpresa;
